Retrieve token manually when fetching URN-based content
