+++ Sources/CoreBusiness/ResourceLoading/AkamaiResourceLoaderDelegate.swift
@@ -7,43 +7,12 @@
 import AVFoundation
 
 final class AkamaiResourceLoaderDelegate: NSObject, AVAssetResourceLoaderDelegate {
-    private static let tokenServiceUrl = URL(string: "https://tp.srgssr.ch/akahd/token")!
     private static let session = URLSession(configuration: .default)
 
     private let id: UUID
 
     init(id: UUID) {
         self.id = id
-    }
-
-    private static func acl(for url: URL) -> String {
-        url.deletingLastPathComponent().appending(component: "*").path
-    }
-
-    private static func tokenRequestUrl(for url: URL) -> URL? {
-        guard var components = URLComponents(url: tokenServiceUrl, resolvingAgainstBaseURL: false) else { return nil }
-        components.queryItems = [
-            URLQueryItem(name: "acl", value: acl(for: url))
-        ]
-        return components.url
-    }
-
-    private static func mergeQueryItems(_ queryItems: [URLQueryItem], into url: URL) -> URL? {
-        guard var components = URLComponents(url: url, resolvingAgainstBaseURL: false) else { return nil }
-        var mergedQueryItems = components.queryItems ?? []
-        mergedQueryItems += queryItems
-        components.queryItems = mergedQueryItems
-        return components.url
-    }
-
-    private static func tokenizeUrl(_ url: URL) async -> URL {
-        guard let requestUrl = tokenRequestUrl(for: url),
-              let (data, _) = try? await session.httpData(from: requestUrl),
-              let tokenPayload = try? JSONDecoder().decode(TokenPayload.self, from: data),
-              let tokenizedUrl = mergeQueryItems(tokenPayload.token.queryItems(), into: url) else {
-            return url
-        }
-        return tokenizedUrl
     }
 
     func resourceLoader(
@@ -61,11 +30,11 @@ final class AkamaiResourceLoaderDelegate: NSObject, AVAssetResourceLoaderDelegat
     }
 
     private func processRequest(_ loadingRequest: AVAssetResourceLoadingRequest) -> Bool {
-        guard let requestUrl = loadingRequest.request.url, let url = AkamaiURLCoding.decodeUrl(requestUrl, id: id) else {
+        guard let requestUrl = loadingRequest.request.url, let url = Akamai.decodeUrl(requestUrl, id: id) else {
             return false
         }
         Task {
-            let tokenizedUrl = await Self.tokenizeUrl(url)
+            let tokenizedUrl = await Akamai.tokenizeUrl(url, using: Self.session)
             loadingRequest.redirect(to: tokenizedUrl)
             loadingRequest.finishLoading()
         }
