+++ Tests/CoreBusinessTests/AkamaiURLCodingTests.swift
@@ -13,13 +13,13 @@ final class AkamaiURLCodingTests: XCTestCase {
     private static let uuid = UUID(uuidString: "E621E1F8-C36C-495A-93FC-0C247A3E6E5F")!
 
     func testEncoding() {
-        expect(AkamaiURLCoding.encodeUrl(
+        expect(Akamai.encodeUrl(
             URL(string: "http://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(equal(URL(string: "akamai+E621E1F8-C36C-495A-93FC-0C247A3E6E5F+http://www.server.com/playlist.m3u8?param1=value1&param2=value2")))
 
-        expect(AkamaiURLCoding.encodeUrl(
+        expect(Akamai.encodeUrl(
             URL(string: "https://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
@@ -27,7 +27,7 @@ final class AkamaiURLCodingTests: XCTestCase {
     }
 
     func testFailedEncoding() {
-        expect(AkamaiURLCoding.encodeUrl(
+        expect(Akamai.encodeUrl(
             URL(string: "//www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
@@ -35,13 +35,13 @@ final class AkamaiURLCodingTests: XCTestCase {
     }
 
     func testDecoding() {
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "akamai+E621E1F8-C36C-495A-93FC-0C247A3E6E5F+http://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(equal(URL(string: "http://www.server.com/playlist.m3u8?param1=value1&param2=value2")))
 
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "akamai+E621E1F8-C36C-495A-93FC-0C247A3E6E5F+https://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
@@ -49,31 +49,31 @@ final class AkamaiURLCodingTests: XCTestCase {
     }
 
     func testFailedDecoding() {
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "http://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(beNil())
 
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "https://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(beNil())
 
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "custom://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(beNil())
 
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "//www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
         .to(beNil())
 
-        expect(AkamaiURLCoding.decodeUrl(
+        expect(Akamai.decodeUrl(
             URL(string: "akamai+1111111-1111-1111-1111-111111111111+https://www.server.com/playlist.m3u8?param1=value1&param2=value2")!,
             id: Self.uuid
         ))
